Fixes type error in test

+++ moveit_planners/pilz_industrial_motion_planner/test/integrationtest_command_list_manager.cpp
@@ -358,7 +358,7 @@ TEST_F(IntegrationTestCommandListManager, blendTwoSegments)
  */
 TEST_F(IntegrationTestCommandListManager, runCallbackForEachItem)
 {
-  int n_items_planned{ 0 };
+  size_t n_items_planned{ 0 };
   ItemPlannedCallback callback{ [&](const planning_interface::MotionPlanResponse&) { ++n_items_planned; } };
   Sequence seq{ data_loader_->getSequence("SimpleSequence") };
   ASSERT_EQ(seq.size(), 2u);
